Tidy build script

+++ build_cmake_pybullet_double.sh
@@ -1,10 +1,16 @@
 #!/bin/sh
-rm CMakeCache.txt
-mkdir build_cmake
+
+if [ -e CMakeCache.txt ]; then
+  rm CMakeCache.txt
+fi
+mkdir -p build_cmake
 cd build_cmake
 cmake -DBUILD_PYBULLET=ON -DBUILD_PYBULLET_NUMPY=OFF -DUSE_DOUBLE_PRECISION=ON -DCMAKE_BUILD_TYPE=Release ..
-make -j12
+make -j $(command nproc || echo 12)
 cd examples
 cd pybullet
-ln -s pybullet.dylib pybullet.so
+if [ -e pybullet.dylib ]; then
+  rm pybullet.so
+  ln -s pybullet.dylib pybullet.so
+fi
